pyproject.toml
[project]
name = "customer-support-assistant"
version = "0.1.0"
description = ""
authors = [
    { name = "stevealila", email = "stevealila25@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic",
    "python-jose",
    "passlib",
    "python-multipart",
    "aiohttp",
    "asyncpg (>=0.30.0,<0.31.0)",
    "groq (>=0.22.0,<0.23.0)"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
